PathToContent causing Invalid file exceptions
Will be adding a few fixes that have known to causes issues with others.  In the event PathToContent had a full path in it instead of just the file name and extension. Added code to extract the file name from the PathToContent and use it instead of the default string form the server