OutOfMemoryException Issue
This has occurred when syncing large volumes of data.  By reading the ehader only, we get what we need up front and let the rest of the content get delivered directly to file reducing the memory footprint needs of the application.

Rearranged function calls in HTTPCall ExecuteAndSaveAsync. Added HttpCompletionOption.ResponseHeadersRead to SendRequestAsync to only read the Http header into memory so we can have the filestreamer read the body directly to file