Auto select config if only one
During the initial sync, it was requested that if there is only one MAC available to selct that the selection window will nto open and the app with automatically select the only MAC and proivide UI accordinly.  In the event of multiple MACs to choose from, the configuratio nselectin UI woudl be presented.
This has been completed